appstore v3.3 push

+++ EazyFitness/DataServer/EFRequest.swift
@@ -14,6 +14,7 @@ class EFRequest: EFData {
     
     var title:String = ""
     var receiver:String = ""
+    var sander:String = ""
     var text:String = ""
     var date:Date = Date()
     var type:requestType = .other
@@ -37,10 +38,11 @@ class EFRequest: EFData {
                 if let data = snap?.data(){
                     self.title = data["title"] as! String
                     self.receiver = data["receiver"] as! String
+                    self.sander = data["sander"] as! String
                     self.text = data["text"] as! String
                     self.date = data["date"] as! Date
-                    self.bageRef = data["bageRef"] as! DocumentReference
-                    self.requestRef = data["requestRef"] as! DocumentReference
+                    self.bageRef = data["bageRef"] as? DocumentReference
+                    self.requestRef = data["requestRef"] as? DocumentReference
                     //studentApproveCourse: studentcourse
                     //trainerApproveCourse: course
                     //studentAddValue:
@@ -52,7 +54,7 @@ class EFRequest: EFData {
         }
     }
     
-    override func upload() {
+    override func upload(handler: (()->())? = nil) {
         
     }
     
@@ -70,12 +72,13 @@ class EFRequest: EFData {
         }
     }
     
-    class func createRequest(bageRef:DocumentReference, title:String, receiver:String, text:String, requestRef:DocumentReference, type:requestType){
+    class func createRequest(bageRef:DocumentReference, title:String, receiver:String, sander:String, text:String, requestRef:DocumentReference, type:requestType){
         Firestore.firestore().collection("request").addDocument(data:[
             "title" : title,
             "receiver" : receiver,
             "text" : text,
             "requestRef" : requestRef,
+            "sander" : sander,
             "type": enumService.toString(e: type),
             "bageRef": bageRef,
             "date": Date()]){ (err) in
@@ -96,7 +99,9 @@ class EFRequest: EFData {
     func cancel(){
         if self.disable == false {
             ref.delete()
-            self.bageRef.delete()
+            if self.bageRef != nil {
+                self.bageRef.delete()
+            }
             switch self.type{
             case .studentAddValue:
                 self.requestRef.delete()
@@ -114,7 +119,15 @@ class EFRequest: EFData {
                 break
             case .other:
                 break
+            case .notification:
+                break
             }
+            
+            if self.type != .notification{
+                let bageRef = Firestore.firestore().collection("Message").addDocument(data: ["uid" : self.sander, "bage":2])
+                EFRequest.createRequest(bageRef: bageRef, title: "\"\(self.title)\"已被拒绝", receiver: self.sander, sander: self.receiver, text: "\"\(self.text)\" 已被拒绝", requestRef: self.ref, type: .notification)
+            }
+            
             self.disable = true
         }
     }
@@ -122,7 +135,9 @@ class EFRequest: EFData {
     func approve(){
         if self.disable == false{
             ref.delete()
-            self.bageRef.delete()
+            if self.bageRef != nil {
+                self.bageRef.delete()
+            }
             switch self.type{
             case .studentAddValue:
                 self.requestRef.updateData(["approved" : true])
@@ -140,13 +155,32 @@ class EFRequest: EFData {
                 break
             case .other:
                 break
+            case .notification:
+                break
             }
+            
+            if self.type != .notification{
+                let bageRef = Firestore.firestore().collection("Message").addDocument(data: ["uid" : self.sander, "bage":2])
+                EFRequest.createRequest(bageRef: bageRef, title: "\"\(self.title)\"已添加成功", receiver: self.sander, sander: self.receiver, text: "\"\(self.text)\" 已添加", requestRef: self.ref, type: .notification)
+            }
+            
             self.disable = true
         }
     }
     
+    func dismiss(){
+        if self.type == .notification{
+            if self.disable == false{
+                ref.delete()
+                if self.bageRef != nil {
+                    self.bageRef.delete()
+                }
+            }
+        }
+    }
+    
     class func getRequestForCurrentUser(type:requestType?){
-        EFRequest.requestList = []
+        print("=================")
         if let currentUserUID = Auth.auth().currentUser?.uid{
             Firestore.firestore().collection("request").whereField("receiver", isEqualTo: currentUserUID).getDocuments { (snap, err) in
                 if let err = err {
@@ -154,12 +188,12 @@ class EFRequest: EFData {
                     message.text = "读取申请时失败: \(err.localizedDescription)"
                     MDCSnackbarManager.show(message)
                 } else {
+                    EFRequest.requestList = []
                     for doc in snap!.documents{
                         if ((enumService.toRequestType(s: doc["type"] as! String) == type && type != nil) || type == nil) {
                             let efRequest = EFRequest(with: doc.reference)
                             efRequest.download()
                             EFRequest.requestList.append(efRequest)
-                            print("EFRequest.requestList.append(efRequest)")
                         }
                     }
                     AppDelegate.load()
